// src/app/model/SortType.ts
export enum SortType {
    IdDesc = 'Id descending',
    IdAsc = 'Id ascending',
    RequestNameDesc = 'Name descending',
    RequestNameAsc = 'Name ascending',
    StatusDesc = 'Status descending',
    StatusAsc = 'Status ascending',
    DeadlineDesc = 'Deadline descending',
    DeadlineAsc = 'Deadline ascending',
    BudgetDesc = 'Budget descending',
    BudgetAsc = 'Budget ascending',
}